Cleanup proxy logic and add routes to server root
