+++ server/middleware/verify-app-proxy.js
@@ -1,0 +1,18 @@
+import { verifyAppProxyExtensionSignature } from "../helpers/app-proxy.js";
+
+// Learn more: https://shopify.dev/apps/online-store/app-proxies#handling-proxy-requests
+export default function verifyAppProxyExtensionSignatureMiddleware(app) {
+  return async (req, res, next) => {
+    if (
+      verifyAppProxyExtensionSignature(
+        req.query,
+        process.env.SHOPIFY_API_SECRET
+      )
+    ) {
+      return next();
+    }
+    res.statusCode = 401;
+  };
+}
+
+//https://github.com/gil--/Store-Recon-Locator/blob/495da3ae3d92250e1e77430f1cf92579e03055e8/server/middleware/verify-proxy.js#L9
